feat: add heading-id link type (#1442)
this adds a new "Heading identifier" type to the `<Link>` widget, which
can be used in tandem with the `<HeadingId>` widget, i.e. when adding a
custom identifier to a heading with the `<HeadingId>` widget, this
identifier can then be used in a `<Link>` widget of type "Heading
identifier" to create in-document links to that headings.