feat: add preview links to cms resources (#1518)
this adds preview links to the cms, which allow live-previewing changes
in a branch on the actual website, without requiring a full rebuild of
the site.

preview mode is enabled for details page of resources (curriculum,
source, event, pathfinder, hosted resource, external resource),
documentation pages, and the home page. it is currently *not* enabled on
any list views (e.g. the resources overview page, but also the resources
list on a source page, or the "related resources" list).

NOTE: images will not be optimized in preview mode, and we don't try to
read image dimensions, so content may jump when images load in (layout
shift).

---

decide how to handle images:
- [ ] try to read the first chunk to actually get correct image
dimensions
- [x] just use `unoptimized: true` or render plain `<img>`
- [ ] patch keystatic image field to store image dimensions on save

decide what to do with list resources (e.g. on `/sources/:id` page):
- [ ] omit
- [x] just show content from `main` branch
- [ ] actually fetch from github

resolves #1425