feat: wrap main content in not-found boundary (#1538)
the non-global not found page will only be used when the `(default)`
route group matches (because that is where `not-found.tsx` lives). this
means `/xxx` will show the global-not-found, but `/sources/xxx` will
show the one wrapped with the default layout.

if we want this not-found page to be shown everywhere, we would need a
catch-all route in the route group, or render the app shell in the
global-not-found.