fix: only enable preview mode with valid branch and token (#1545)
when draft mode is enabled (i.e. we have a `__prerender_bypass` cookie)
but don't have a valid branch name or token (read from cookies), we
should not enable preview mode. we still want to render the orange
banner though, so a user can click Disable. we also call
`draftMode.disable` in that case, which does the same as clicking
Disable, and surprisingly seems to work even though it is not called in
a route handler.