Content/litriddle (#1508)
Hi Vicky, 

I have tweaked my content based on a test user's (my poor husband)
feedback. There are some slight differences between the webpage here,
and the R markdown file that I will email you separately, as it's not a
supported file type here. The differences are mainly in how I refer to
the navigation (like; the green button for running code, which is of
course not present on the webpage).
I'm a bit unsure about including the exercises on the webpage.
Theoretically people can just type out the code in a notepad or even on
paper, but I wonder if people will actually do that.
Any other feedback is of course very welcome!

Cheers,
Floor

---------

Co-authored-by: stefanprobst <stefanprobst@posteo.de>
Co-authored-by: VickyGarnett <57193002+VickyGarnett@users.noreply.github.com>